Бот теперь общается с каждым пользователем индивидуально, включая показ списка покупок, исправил файловый набор и набор директорий, убрал лишние хендлеры (из трека)

+++ shopping_manager_bot/db/CRUD.py
@@ -7,18 +7,22 @@ def add_one_good(user_id: int, good: str) -> None:
     db_session.commit()
 
 
-def get_all_goods() -> list | None:
-    content = Good.query.all()
-    return content
+def get_all_goods(user_id: int) -> list | None:
+    return Good.query.filter(Good.user_id == user_id).all()
 
 
-def delete_all_goods() -> None:
-    Good.query.delete()
+def delete_all_goods(user_id: int) -> None:
+    Good.query.filter(Good.user_id == user_id).delete()
     db_session.commit()
 
 
-def delete_one_good(content_id: int) -> None:
-    content = Good.query.get(content_id)
-    db_session.delete(content)
-    db_session.commit()
+def delete_one_good(content_id: int, user_id: int) -> bool:
+    result = Good.query.with_entities(Good.id).filter(Good.user_id == user_id).all()
+    for good_id in result:
+        if content_id == good_id[0]:
+            content = Good.query.get(content_id)
+            db_session.delete(content)
+            db_session.commit()
+            return True
+    return None
